+++ posthog-node/src/feature-flags.ts
@@ -190,10 +190,34 @@ class FeatureFlagsPoller {
     let isInconclusive = false
     let result = undefined
 
-    flagConditions.forEach((condition) => {
+    // # Stable sort conditions with variant overrides to the top. This ensures that if overrides are present, they are
+    // # evaluated first, and the variant override is applied to the first matching condition.
+    const sortedFlagConditions = [...flagConditions].sort((conditionA, conditionB) => {
+      const AHasVariantOverride = !!conditionA.variant
+      const BHasVariantOverride = !!conditionB.variant
+
+      if (AHasVariantOverride && BHasVariantOverride) {
+        return 0
+      } else if (AHasVariantOverride) {
+        return -1
+      } else if (BHasVariantOverride) {
+        return 1
+      } else {
+        return 0
+      }
+    })
+
+    for (const condition of sortedFlagConditions) {
       try {
         if (this.isConditionMatch(flag, distinctId, condition, properties)) {
-          result = this.getMatchingVariant(flag, distinctId) || true
+          const variantOverride = condition.variant
+          const flagVariants = flagFilters.multivariate?.variants || []
+          if (variantOverride && flagVariants.some((variant) => variant.key === variantOverride)) {
+            result = variantOverride
+          } else {
+            result = this.getMatchingVariant(flag, distinctId) || true
+          }
+          break
         }
       } catch (e) {
         if (e instanceof InconclusiveMatchError) {
@@ -202,7 +226,7 @@ class FeatureFlagsPoller {
           throw e
         }
       }
-    })
+    }
 
     if (result !== undefined) {
       return result